+++ plasma_framework/contracts/mocks/framework/utils/ExitPriorityWrapper.sol
@@ -4,7 +4,7 @@ import "../../../src/framework/utils/ExitPriority.sol";
 import "../../../src/utils/PosLib.sol";
 
 contract ExitPriorityWrapper {
-    function computePriority(uint64 exitableAt, uint256 txPos, uint160 exitId) public pure returns (uint256) {
+    function computePriority(uint64 exitableAt, uint256 txPos, uint168 exitId) public pure returns (uint256) {
         return ExitPriority.computePriority(exitableAt, PosLib.decode(txPos), exitId);
     }
 
@@ -12,7 +12,7 @@ contract ExitPriorityWrapper {
         return ExitPriority.parseExitableAt(priority);
     }
 
-    function parseExitId(uint256 priority) public pure returns (uint160) {
+    function parseExitId(uint256 priority) public pure returns (uint168) {
         return ExitPriority.parseExitId(priority);
     }
 }
